// apps/dashboard/src/app/(app)/team/[team_slug]/[project_slug]/(sidebar)/ai/api/types.ts
type SessionContextFilter = {
  chain_ids: string[] | null;
  wallet_address: string | null;
};

type NebulaUserMessageContentItem =
  | {
      type: "image";
      image_url: string | null;
      b64: string | null;
    }
  | {
      type: "text";
      text: string;
    }
  | {
      type: "transaction";
      transaction_hash: string;
      chain_id: number;
    };

export type NebulaUserMessageContent = NebulaUserMessageContentItem[];

export type NebulaUserMessage = {
  role: "user";
  content: NebulaUserMessageContent;
};

export type NebulaSessionHistoryMessage =
  | {
      role: "assistant" | "action" | "image";
      content: string;
      timestamp: number;
    }
  | {
      role: "user";
      content: NebulaUserMessageContent | string;
    };

export type SessionInfo = {
  id: string;
  account_id: string;
  model_name: string;
  archive_at: string | null;
  can_execute: boolean;
  created_at: string;
  deleted_at: string | null;
  history: Array<NebulaSessionHistoryMessage> | null;
  updated_at: string;
  archived_at: string | null;
  title: string | null;
  is_public: boolean | null;
  context: SessionContextFilter | null;
  // memory
  // action: array<object> | null; <-- type of this is not available on https://nebula-api.thirdweb-dev.com/docs#/default/get_session_session__session_id__get
};

export type UpdatedSessionInfo = {
  title: string;
  model_name: string;
  account_id: string;
  context: SessionContextFilter | null;
};

export type DeletedSessionInfo = {
  id: string;
  deleted_at: string;
};

export type TruncatedSessionInfo = {
  created_at: string;
  id: string;
  updated_at: string;
  title: string | null;
};

export type NebulaTxData = {
  chain_id: number;
  data: `0x${string}`;
  to: string;
  value?: string;
};

export type NebulaContext = {
  chainIds: string[] | null;
  walletAddress: string | null;
  networks: "mainnet" | "testnet" | "all" | null;
};

export type NebulaSwapData = {
  action: string;
  transaction: {
    chain_id: number;
    to: `0x${string}`;
    data: `0x${string}`;
    value?: string;
  };
  to_token: {
    address: `0x${string}`;
    amount: string;
    chain_id: number;
    decimals: number;
    symbol: string;
  };
  from_token: {
    address: `0x${string}`;
    amount: string;
    chain_id: number;
    decimals: number;
    symbol: string;
  };
  intent: {
    amount: string;
    destination_chain_id: number;
    destination_token_address: `0x${string}`;
    origin_chain_id: number;
    origin_token_address: `0x${string}`;
    receiver: `0x${string}`;
    sender: `0x${string}`;
  };
};